+++ internal/framework/services/compute/data_sources.go
@@ -1,0 +1,12 @@
+package compute
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+)
+
+// DataSources returns the list of data sources for the package.
+func DataSources() []func() datasource.DataSource {
+	return []func() datasource.DataSource{
+		NewContractsDataSource,
+	}
+}